gpuas-test-framework/requirements.txt
# Testing Framework
pytest==7.4.0
pytest-html==4.0.2
pytest-xdist==3.3.1
pytest-asyncio==0.21.0
allure-pytest==2.13.2
pytest-playwright==0.4.3

# API Testing
requests==2.31.0
httpx==0.24.1
pydantic==2.12.5

# UI Testing
playwright==1.39.0

# Performance Testing
locust==2.20.0

# Mock Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Utilities
python-dotenv==1.0.0
factory-boy==3.3.0
Faker==20.1.0
pandas==2.2.3
numpy==2.2.6

# Compatibility override
greenlet==3.3.0

# DevOps
docker==6.1.3
python-jenkins==1.7.0
